+++ ui/tauri-gtk-webview.h
@@ -1,0 +1,14 @@
+#include <JavaScriptCore/JavaScript.h>
+#include <gtk/gtk.h>
+#include <webkit2/webkit2.h>
+
+struct webview_priv {
+  GtkWidget *window;
+  GtkWidget *scroller;
+  GtkWidget *webview;
+  GtkWidget *inspector_window;
+  GAsyncQueue *queue;
+  int ready;
+  int js_busy;
+  int should_exit;
+};
