+++ ui/tauri-windows-webview.h
@@ -1,0 +1,19 @@
+#define CINTERFACE
+#include <windows.h>
+
+#include <commctrl.h>
+#include <exdisp.h>
+#include <mshtmhst.h>
+#include <mshtml.h>
+#include <shobjidl.h>
+
+#include <stdio.h>
+
+struct webview_priv {
+  HWND hwnd;
+  IOleObject **browser;
+  BOOL is_fullscreen;
+  DWORD saved_style;
+  DWORD saved_ex_style;
+  RECT saved_rect;
+};
